Added tests for FlacFileStream and split it from __init__.py
